Move keyboard definitions to new file
Also add a function to align keyboard on toggle
